基本機能の実装完了

+++ src/main/java/org/example/Main.java
@@ -1,56 +1,35 @@
 package org.example;
 
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.core.util.JacksonFeatureSet;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.poi.ss.usermodel.*;
-import org.apache.poi.ss.util.CellRangeAddress;
-import org.apache.poi.ss.util.CellReference;
-import org.apache.poi.xssf.usermodel.XSSFName;
-import org.apache.poi.xssf.usermodel.XSSFRow;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URL;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
+import java.io.*;
+import java.net.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.security.CodeSource;
+import java.security.ProtectionDomain;
 
 public class Main {
-    public static void main(String[] args) throws IOException, InterruptedException {
-        System.out.println("Hello world!");
-        Form form = new Form(args[0]);
-//        form.setValues();
-        form.setRowValues();
-        form.writeFile();
-
-//// create a client
-//        var client = HttpClient.newHttpClient();
-//
-//// create a request
-//        var request = HttpRequest.newBuilder(
-//                        URI.create("https://api.nasa.gov/planetary/apod?api_key=DEMO_KEY"))
-//                .header("accept", "application/json")
-//                .build();
-//
-//// use the client to send the request
-////        var response = client.send(request, new JsonBodyHandler<>(APOD.class));
-//        var response = client.send(request, HttpResponse.BodyHandlers.ofString());
-//        ObjectMapper mapper = new ObjectMapper();
-//        var json = "{\"key\": {\"value\": \"123\"}}";
-//        Map<String, Object> map = mapper.readValue(json, HashMap.class);
-//        System.out.println(map.get("key"));
-//// the response:®
-//        System.out.println(map);
+    public static void main(String[] args) throws Exception {
+        if (args.length != 3) {
+            throw new Exception("パラメータの数が違います。[0]ユーザーID, [1]ファイル名, [2]データ（json）を渡してください。");
+        }
+        Path path = getApplicationPath(Main.class).getParent().resolve("files/" + args[0] + "/" + args[1]);
+        String filePathString = path.toString() + ".xlsx";
+        if (!Files.exists(Paths.get(filePathString))) {
+            throw  new FileNotFoundException("対象のファイルが存在しません。" + filePathString);
+        }
+        String json = args[2];
+        Form form = new Form(filePathString, json);
+        form.setValues();
+//        form.setRowValues();
+        String base64 = form.writeFile();
+        System.out.println(base64);
+    }
+    public static Path getApplicationPath(Class<?> cls) throws URISyntaxException {
+        ProtectionDomain pd = cls.getProtectionDomain();
+        CodeSource cs = pd.getCodeSource();
+        URL location = cs.getLocation();
+        URI uri = location.toURI();
+        return Paths.get(uri);
     }
 }
